try to fix the world.

+++ Integrations/EWSMailSender/EWSMailSender.py
@@ -317,7 +317,16 @@ def main():
     finally:
         if isinstance(config, Configuration):
             # The protocol will not kill its threads after use, so kill it manually
-            config.protocol.thread_pool.terminate()
+            if "thread_pool" in config.protocol.__dict__:
+                config.protocol.thread_pool.terminate()
+                del config.protocol.__dict__["thread_pool"]
+        try:
+            # we don't want to leave cached connection around as EWS limits the number of connections
+            # in a very aggressive way. 12 seems to be the default limit
+            # see: https://blogs.msdn.microsoft.com/webdav_101/2018/06/02/you-are-doing-too-much-at-one-time-ewsmaxconcurrency-too-many-concurrent-connections-opened/ # noqa
+            close_connections()
+        except Exception as ex:
+            demisto.info("Failed close_connections (shouldn't happen). Ignoring exception: {}".format(ex))
         if log_stream:
             try:
                 logging.getLogger().removeHandler(log_handler)  # type: ignore
